remove all apis, add user level example
